refactor : 함수들과 변수들의 파일 위치 변경
- webtoon data를 webtoon.js에서 data.js 파일로 이동
- renderCarousel 함수를 webtoon.js에서 carousel.js로 이동
